refactor: 금액 정수 수정

+++ src/main/java/kr/hhplus/be/server/common/event/balance/BalanceDeductedEvent.java
@@ -2,6 +2,6 @@ package kr.hhplus.be.server.common.event.balance;
 
 public record BalanceDeductedEvent(
         Long orderId,
-        Long totalAmount
+        Integer totalAmount
 ) {
 }
