+++ src/main/java/kr/hhplus/be/server/common/event/coupon/CouponUsedEvent.java
@@ -2,7 +2,7 @@ package kr.hhplus.be.server.common.event.coupon;
 
 public record CouponUsedEvent(
         Long orderId,
-        Long discountAmount,
+        Integer discountAmount,
         Long couponId
 ) {
 }
