+++ src/main/java/kr/hhplus/be/server/common/event/product/StockReservedEvent.java
@@ -2,6 +2,6 @@ package kr.hhplus.be.server.common.event.product;
 
 public record StockReservedEvent(
         Long orderId,
-        Long subTotalAmount
+        Integer subTotalAmount
 ) {
 }
